+++ api/shared-modules/src/web/models/schemas/index.ts
@@ -14,6 +14,7 @@ export {
 	IdentityKeysSchema,
 	IdentityKeyPairSchema,
 	VerifiableCredentialSchema,
+	VerifiablePresentationSchema,
 	VerifiableCredentialSubjectSchema
 } from './identity';
 export { AccessRights, SubscriptionSchema, SubscriptionType, SubscriptionStateSchema, SubscriptionUpdateSchema } from './subscription';