+++ api/shared-modules/src/web/models/schemas/request-response-body/index.ts
@@ -23,6 +23,5 @@ export {
 	CreateCredentialBodySchema,
 	RevokeVerificationBodySchema,
 	SubjectBodySchema,
-	TrustedRootBodySchema,
-	VerifiableCredentialBodySchema
+	TrustedRootBodySchema
 } from './verification-bodies';