+++ api/ssi-bridge/src/services/verification-service.ts
@@ -5,7 +5,8 @@ import {
 	Credential,
 	IdentityKeys,
 	Subject,
-	Bitmap
+	Bitmap,
+	VerifiablePresentation
 } from '@iota/is-shared-modules';
 import { SsiService } from './ssi-service';
 import { UserService } from './user-service';
@@ -98,6 +99,28 @@ export class VerificationService {
 
 		const isTrustedIssuer = trustedRoots && trustedRoots.some((rootId) => rootId === vc.issuer);
 		const isVerified = isVerifiedCredential && isTrustedIssuer;
+		return isVerified;
+	}
+
+	async checkVerifiablePresentation(vp: VerifiablePresentation): Promise<boolean> {
+		const expiration = 60;
+		const isVerifiedVP = await this.ssiService.checkVerifiablePresentation(vp, expiration);
+		const trustedRoots = await this.getTrustedRootIds();
+		let isTrustedIssuer = false;
+
+		if (!trustedRoots) {
+			throw Error('No trusted root ids found!');
+		}
+
+		if (Array.isArray(vp.verifiableCredential)) {
+			isTrustedIssuer = trustedRoots.some((rootId) =>
+				(vp.verifiableCredential as VerifiableCredential[]).some((vc) => vc.issuer == rootId)
+			);
+		} else {
+			isTrustedIssuer = trustedRoots.some((rootId) => rootId === (vp.verifiableCredential as VerifiableCredential).issuer);
+		}
+
+		const isVerified = isVerifiedVP && isTrustedIssuer;
 		return isVerified;
 	}