+++ clients/client-sdk/src/models/clientConfig.ts
@@ -1,9 +1,11 @@
+import { ApiVersion } from './apiVersion';
+
 export type ClientConfig = {
   apiKey?: string;
   isGatewayUrl?: string;
   useGatewayUrl?: boolean;
   ssiBridgeUrl?: string;
   auditTrailUrl?: string;
-  apiVersionAuditTrail: string;
-  apiVersionSsiBridge: string
+  apiVersionAuditTrail: ApiVersion;
+  apiVersionSsiBridge: ApiVersion;
 };